feat(fixture): separate module for fixtures
- we want to export fixtures so that other repositories can use them the
way this service intends for them to be used. indexd_server and indexd_client
are the two main fixtures.
